fix(types): move Express Request augmentation to shared types file
- Rename express.d.ts to express.ts for proper module resolution
- Move auditContext declaration from logger.middleware.ts to shared types
- Add import for Express augmentation in exception filters
