+++ apps/backend/src/common/exceptions/all-exceptions.filter.ts
@@ -13,6 +13,8 @@ import {
   GENERIC_ERROR_MESSAGES,
 } from './error-sanitizer';
 import { isProduction } from 'src/config/environment.config';
+// Import for Express Request augmentation (auditContext)
+import '../types/express';
 
 /**
  * Global Exception Filter
