+++ apps/backend/src/common/exceptions/http-exception.filter.ts
@@ -7,6 +7,8 @@ import {
 } from '@nestjs/common';
 import { Request, Response } from 'express';
 import { sanitizeErrorMessage, logErrorDetails } from './error-sanitizer';
+// Import for Express Request augmentation (auditContext)
+import '../types/express';
 
 /**
  * HTTP Exception Filter
